+++ admin-console/src/main/java/io/information/modules/news/entity/CardBaseEntity.java
@@ -45,7 +45,7 @@ public class CardBaseEntity implements Serializable {
     /**
      * 帖子节点分类（字典）
      */
-    private Integer cNodeCategory;
+    private Long cNodeCategory;
     /**
      * 帖子标题
      */
@@ -129,11 +129,11 @@ public class CardBaseEntity implements Serializable {
         this.cCategory = cCategory;
     }
 
-    public Integer getcNodeCategory() {
+    public Long getcNodeCategory() {
         return cNodeCategory;
     }
 
-    public void setcNodeCategory(Integer cNodeCategory) {
+    public void setcNodeCategory(Long cNodeCategory) {
         this.cNodeCategory = cNodeCategory;
     }
